Config parameters for integration test timing
